+++ package.json
@@ -23,7 +23,6 @@
     "tailwindcss-capsize": "^3.0.2"
   },
   "dependencies": {
-    "@legendapp/state": "^0.19.0",
     "@nanostores/react": "^0.4.0",
     "@react-three/drei": "^9.32.7",
     "@react-three/fiber": "^8.8.8",
